requirements.txt
streamlit
pandas
psycopg2-binary
sqlalchemy
streamlit-aggrid
pyotp
qrcode
bcrypt
cryptography
pillow
supabase
